+++ pennylane_lightning/lightning_qubit.py
@@ -15,6 +15,7 @@ r"""
 This module contains the :class:`~.LightningQubit` class, a PennyLane simulator device that
 interfaces with C++ for fast linear algebra calculations.
 """
+from typing import List
 from warnings import warn
 
 import numpy as np
@@ -29,14 +30,17 @@ from pennylane import (
 import pennylane as qml
 from pennylane.devices import DefaultQubit
 from pennylane.operation import Expectation
+from pennylane.wires import Wires
 
 from ._version import __version__
 
 try:
     from .lightning_qubit_ops import (
+        MeasuresC64,
         StateVectorC64,
         AdjointJacobianC64,
         VectorJacobianProductC64,
+        MeasuresC128,
         StateVectorC128,
         AdjointJacobianC128,
         VectorJacobianProductC128,
@@ -302,108 +306,20 @@ class LightningQubit(DefaultQubit):
             tp_shift,
             tape.num_params,
         )
-        return jac
-
-    def vector_jacobian_product(self, tape, dy, starting_state=None, use_device_state=False):
-        """Generate the the vector-Jacobian products of a tape.
-
-        Args:
-            tape (.QuantumTape): quantum tape to differentiate
-            dy (tensor_like): Gradient-output vector. Must have shape
-                matching the output shape of the corresponding tape.
-
-        Keyword Args:
-            starting_state (tensor_like): post-forward pass state to start execution with. It should be
-                complex-valued. Takes precedence over ``use_device_state``.
-            use_device_state (bool): use current device state to initialize. A forward pass of the same
-                circuit should be the last thing the device has executed. If a ``starting_state`` is
-                provided, that takes precedence.
-
-        Returns:
-            tuple[array or None, tensor_like or None]: A tuple of the adjoint-jacobian and the Vector-Jacobian
-            product. Returns ``None`` if the tape has no trainable parameters.
-        """
-        if self.shots is not None:
-            warn(
-                "Requested adjoint differentiation to be computed with finite shots."
-                " The derivative is always exact when using the adjoint differentiation method.",
-                UserWarning,
-            )
-
-        num_params = len(tape.trainable_params)
-
-        if num_params == 0:
-            return None, None
-
-        if math.allclose(dy, 0):
-            return None, math.convert_like(np.zeros([num_params]), dy)
-
-        # To support np.complex64 based on the type of self._state
-        dtype = self._state.dtype
-        if dtype == np.complex64:
-            use_csingle = True
-        elif dtype == np.complex128:
-            use_csingle = False
-        else:
-            raise TypeError(f"Unsupported complex Type: {dtype}")
-
-        # Check adjoint diff support
-        self.adjoint_diff_support_check(tape)
-
-        # Initialization of state
-        if starting_state is not None:
-            ket = np.ravel(starting_state)
-        else:
-            if not use_device_state:
-                self.reset()
-                self.execute(tape)
-            ket = np.ravel(self._pre_rotated_state)
-
-        if use_csingle:
-            VJP = VectorJacobianProductC64()
-            ket = ket.astype(np.complex64)
-        else:
-            VJP = VectorJacobianProductC128()
-
-        obs_serialized = _serialize_obs(tape, self.wire_map, use_csingle=use_csingle)
-        ops_serialized, use_sp = _serialize_ops(tape, self.wire_map, use_csingle=use_csingle)
-
-        ops_serialized = VJP.create_ops_list(*ops_serialized)
-
-        trainable_params = sorted(tape.trainable_params)
-        first_elem = 1 if trainable_params[0] == 0 else 0
-
-        tp_shift = (
-            trainable_params if not use_sp else [i - 1 for i in trainable_params[first_elem:]]
-        )  # exclude first index if explicitly setting sv
-
-        state_vector = StateVectorC64(ket) if use_csingle else StateVectorC128(ket)
-
-        jac, vjp = VJP.vjp(
-            math.reshape(dy, [-1]),
-            state_vector,
-            obs_serialized,
-            ops_serialized,
-            tp_shift,
-            tape.num_params,
-        )
-        return jac, vjp
+        return jac.reshape(-1, tape.num_params)
 
     def compute_vjp(self, dy, jac, num=None):
         """Convenience function to compute the vector-Jacobian product for a given
         vector of gradient outputs and a Jacobian.
-
         Args:
             dy (tensor_like): vector of gradient outputs
             jac (tensor_like): Jacobian matrix. For an n-dimensional ``dy``
                 vector, the first n-dimensions of ``jac`` should match
                 the shape of ``dy``.
-
         Keyword Args:
         num (int): The length of the flattened ``dy`` argument. This is an
             optional argument, but can be useful to provide if ``dy`` potentially
             has no shape (for example, due to tracing or just-in-time compilation).
-
         Returns:
             tensor_like: the vector-Jacobian product
         """
@@ -441,17 +357,94 @@ class LightningQubit(DefaultQubit):
         )
         return vjp_tensor
 
+    def vjp(self, tape, dy, starting_state=None, use_device_state=False):
+        """Generate the processing function required to compute the vector-Jacobian products of a tape.
+        Args:
+            tape (.QuantumTape): quantum tape to differentiate
+            dy (tensor_like): Gradient-output vector. Must have shape
+                matching the output shape of the corresponding tape.
+        Keyword Args:
+            starting_state (tensor_like): post-forward pass state to start execution with. It should be
+                complex-valued. Takes precedence over ``use_device_state``.
+            use_device_state (bool): use current device state to initialize. A forward pass of the same
+                circuit should be the last thing the device has executed. If a ``starting_state`` is
+                provided, that takes precedence.
+        Returns:
+            The processing function required to compute the vector-Jacobian
+            products of a tape.
+        """
+        if self.shots is not None:
+            warn(
+                "Requested adjoint differentiation to be computed with finite shots."
+                " The derivative is always exact when using the adjoint differentiation method.",
+                UserWarning,
+            )
+
+        num_params = len(tape.trainable_params)
+
+        if num_params == 0:
+            return lambda _: None
+
+        if math.allclose(dy, 0):
+            return lambda _: math.convert_like(np.zeros([num_params]), dy)
+
+        # To support np.complex64 based on the type of self._state
+        dtype = self._state.dtype
+        if dtype == np.complex64:
+            use_csingle = True
+        elif dtype == np.complex128:
+            use_csingle = False
+        else:
+            raise TypeError(f"Unsupported complex Type: {dtype}")
+
+        V = VectorJacobianProductC64() if use_csingle else VectorJacobianProductC128()
+
+        fn = V.vjp_fn(math.reshape(dy, [-1]), tape.num_params)
+
+        def processing_fn(tape):
+            # Check adjoint diff support
+            self.adjoint_diff_support_check(tape)
+
+            # Initialization of state
+            if starting_state is not None:
+                ket = np.ravel(starting_state)
+            else:
+                if not use_device_state:
+                    self.reset()
+                    self.execute(tape)
+                ket = np.ravel(self._pre_rotated_state)
+
+            if use_csingle:
+                ket = ket.astype(np.complex64)
+
+            obs_serialized = _serialize_obs(tape, self.wire_map, use_csingle=use_csingle)
+            ops_serialized, use_sp = _serialize_ops(tape, self.wire_map, use_csingle=use_csingle)
+
+            ops_serialized = V.create_ops_list(*ops_serialized)
+
+            trainable_params = sorted(tape.trainable_params)
+            first_elem = 1 if trainable_params[0] == 0 else 0
+
+            tp_shift = (
+                trainable_params if not use_sp else [i - 1 for i in trainable_params[first_elem:]]
+            )  # exclude first index if explicitly setting sv
+
+            state_vector = StateVectorC64(ket) if use_csingle else StateVectorC128(ket)
+
+            return fn(state_vector, obs_serialized, ops_serialized, tp_shift)
+
+        return processing_fn
+
     def batch_vjp(
         self, tapes, dys, reduction="append", starting_state=None, use_device_state=False
     ):
-        """Generate the the vector-Jacobian products of a batch of tapes.
-
+        """Generate the processing function required to compute the vector-Jacobian products
+        of a batch of tapes.
         Args:
             tapes (Sequence[.QuantumTape]): sequence of quantum tapes to differentiate
             dys (Sequence[tensor_like]): Sequence of gradient-output vectors ``dy``. Must be the
                 same length as ``tapes``. Each ``dy`` tensor should have shape
                 matching the output shape of the corresponding tape.
-
         Keyword Args:
             reduction (str): Determines how the vector-Jacobian products are returned.
                 If ``append``, then the output of the function will be of the form
@@ -462,36 +455,187 @@ class LightningQubit(DefaultQubit):
             use_device_state (bool): use current device state to initialize. A forward pass of the same
                 circuit should be the last thing the device has executed. If a ``starting_state`` is
                 provided, that takes precedence.
-
         Returns:
-            tuple[List[array or None], List[tensor_like or None]]: A tuple containing a list
-            of adjoint-jacobians and a list of vector-Jacobian products. ``None`` elements corresponds
-            to tapes with no trainable parameters.
+            The processing function required to compute the vector-Jacobian products of a batch of tapes.
         """
-        vjps = []
-        jacs = []
+        fns = []
 
         # Loop through the tapes and dys vector
         for tape, dy in zip(tapes, dys):
-            jac, vjp = self.vector_jacobian_product(
-                tape,
-                dy,
-                starting_state=starting_state,
-                use_device_state=use_device_state,
+            fn = self.vjp(
+                tape, dy, starting_state=starting_state, use_device_state=use_device_state
             )
-            if vjp is None:
-                if reduction == "append":
-                    vjps.append(None)
-                    jacs.append(jac)
-                continue
-            if isinstance(reduction, str):
-                getattr(vjps, reduction)(vjp)
-                getattr(jacs, reduction)(jac)
-            elif callable(reduction):
-                reduction(vjps, vjp)
-                reduction(jacs, jac)
+            fns.append(fn)
 
-        return jacs, vjps
+        def processing_fns(tapes):
+            vjps = []
+            for t, f in zip(tapes, fns):
+                vjp = f(t)
+
+                if vjp is None:
+                    if reduction == "append":
+                        vjps.append(None)
+                    continue
+
+                if isinstance(reduction, str):
+                    getattr(vjps, reduction)(vjp)
+                elif callable(reduction):
+                    reduction(vjps, vjp)
+
+            return vjps
+
+        return processing_fns
+
+    def probability(self, wires=None, shot_range=None, bin_size=None):
+        """Return the probability of each computational basis state.
+
+        Devices that require a finite number of shots always return the
+        estimated probability.
+
+        Args:
+            wires (Iterable[Number, str], Number, str, Wires): wires to return
+                marginal probabilities for. Wires not provided are traced out of the system.
+            shot_range (tuple[int]): 2-tuple of integers specifying the range of samples
+                to use. If not specified, all samples are used.
+            bin_size (int): Divides the shot range into bins of size ``bin_size``, and
+                returns the measurement statistic separately over each bin. If not
+                provided, the entire shot range is treated as a single bin.
+
+        Returns:
+            array[float]: list of the probabilities
+        """
+        if self.shots is not None:
+            return self.estimate_probability(wires=wires, shot_range=shot_range, bin_size=bin_size)
+
+        wires = wires or self.wires
+        wires = Wires(wires)
+
+        # translate to wire labels used by device
+        device_wires = self.map_wires(wires)
+
+        # To support np.complex64 based on the type of self._state
+        dtype = self._state.dtype
+        if dtype == np.complex64:
+            use_csingle = True
+        elif dtype == np.complex128:
+            use_csingle = False
+        else:
+            raise TypeError(f"Unsupported complex Type: {dtype}")
+
+        # Initialization of state
+        ket = np.ravel(self._state)
+
+        if use_csingle:
+            ket = ket.astype(np.complex64)
+
+        state_vector = StateVectorC64(ket) if use_csingle else StateVectorC128(ket)
+        M = MeasuresC64(state_vector) if use_csingle else MeasuresC128(state_vector)
+
+        return M.probs(device_wires)
+
+    def expval(self, observable, shot_range=None, bin_size=None):
+        """Expectation value of the supplied observable.
+
+        Args:
+            observable: A PennyLane observable.
+            shot_range (tuple[int]): 2-tuple of integers specifying the range of samples
+                to use. If not specified, all samples are used.
+            bin_size (int): Divides the shot range into bins of size ``bin_size``, and
+                returns the measurement statistic separately over each bin. If not
+                provided, the entire shot range is treated as a single bin.
+
+        Returns:
+            Expectation value of the observable
+        """
+        if isinstance(observable.name, List) or observable.name in [
+            "Identity",
+            "Projector",
+            "Hermitian",
+            "Hamiltonian",
+            "SparseHamiltonian",
+        ]:
+            # TODO: requires backend support
+            return super().expval(observable, shot_range=shot_range, bin_size=bin_size)
+
+        if self.shots is not None:
+            # estimate the expectation value
+            # TODO: Lightning support for sampling
+            samples = self.sample(observable, shot_range=shot_range, bin_size=bin_size)
+            return np.squeeze(np.mean(samples, axis=0))
+
+        # To support np.complex64 based on the type of self._state
+        dtype = self._state.dtype
+        if dtype == np.complex64:
+            use_csingle = True
+        elif dtype == np.complex128:
+            use_csingle = False
+        else:
+            raise TypeError(f"Unsupported complex Type: {dtype}")
+
+        # Initialization of state
+        ket = np.ravel(self._pre_rotated_state)
+
+        if use_csingle:
+            ket = ket.astype(np.complex64)
+
+        state_vector = StateVectorC64(ket) if use_csingle else StateVectorC128(ket)
+        M = MeasuresC64(state_vector) if use_csingle else MeasuresC128(state_vector)
+
+        # translate to wire labels used by device
+        observable_wires = self.map_wires(observable.wires)
+
+        return M.expval(observable.name, observable_wires)
+
+    def var(self, observable, shot_range=None, bin_size=None):
+        """Variance of the supplied observable.
+
+        Args:
+            observable: A PennyLane observable.
+            shot_range (tuple[int]): 2-tuple of integers specifying the range of samples
+                to use. If not specified, all samples are used.
+            bin_size (int): Divides the shot range into bins of size ``bin_size``, and
+                returns the measurement statistic separately over each bin. If not
+                provided, the entire shot range is treated as a single bin.
+
+        Returns:
+            Variance of the observable
+        """
+        if isinstance(observable.name, List) or observable.name in [
+            "Identity",
+            "Projector",
+            "Hermitian",
+        ]:
+            # TODO: requires backend support
+            return super().var(observable, shot_range=shot_range, bin_size=bin_size)
+
+        if self.shots is not None:
+            # estimate the var
+            # TODO: Lightning support for sampling
+            samples = self.sample(observable, shot_range=shot_range, bin_size=bin_size)
+            return np.squeeze(np.var(samples, axis=0))
+
+        # To support np.complex64 based on the type of self._state
+        dtype = self._state.dtype
+        if dtype == np.complex64:
+            use_csingle = True
+        elif dtype == np.complex128:
+            use_csingle = False
+        else:
+            raise TypeError(f"Unsupported complex Type: {dtype}")
+
+        # Initialization of state
+        ket = np.ravel(self._pre_rotated_state)
+
+        if use_csingle:
+            ket = ket.astype(np.complex64)
+
+        state_vector = StateVectorC64(ket) if use_csingle else StateVectorC128(ket)
+        M = MeasuresC64(state_vector) if use_csingle else MeasuresC128(state_vector)
+
+        # translate to wire labels used by device
+        observable_wires = self.map_wires(observable.wires)
+
+        return M.var(observable.name, observable_wires)
 
 
 if not CPP_BINARY_AVAILABLE: