+++ pennylane_lightning/src/algorithms/AdjointDiff.hpp
@@ -24,8 +24,9 @@
 #include <vector>
 
 #include "Error.hpp"
+#include "JacobianTape.hpp"
+#include "LinearAlgebra.hpp"
 #include "StateVectorManaged.hpp"
-#include "Util.hpp"
 
 #include <iostream>
 
@@ -98,230 +99,6 @@ void applyGeneratorControlledPhaseShift(SVType &sv,
 /// @endcond
 
 namespace Pennylane::Algorithms {
-
-/**
- * @brief Utility struct for observable operations used by AdjointJacobian
- * class.
- *
- */
-template <class T = double> class ObsDatum {
-  public:
-    /**
-     * @brief Variant type of stored parameter data.
-     */
-    using param_var_t = std::variant<std::monostate, std::vector<T>,
-                                     std::vector<std::complex<T>>>;
-
-    /**
-     * @brief Copy constructor for an ObsDatum object, representing a given
-     * observable.
-     *
-     * @param obs_name Name of each operation of the observable. Tensor product
-     * observables have more than one operation.
-     * @param obs_params Parameters for a given obserable operation ({} if
-     * optional).
-     * @param obs_wires Wires upon which to apply operation. Each observable
-     * operation will be a separate nested list.
-     */
-    ObsDatum(std::vector<std::string> obs_name,
-             std::vector<param_var_t> obs_params,
-             std::vector<std::vector<size_t>> obs_wires)
-        : obs_name_{std::move(obs_name)},
-          obs_params_(std::move(obs_params)), obs_wires_{
-                                                  std::move(obs_wires)} {};
-
-    /**
-     * @brief Get the number of operations in observable.
-     *
-     * @return size_t
-     */
-    [[nodiscard]] auto getSize() const -> size_t { return obs_name_.size(); }
-    /**
-     * @brief Get the name of the observable operations.
-     *
-     * @return const std::vector<std::string>&
-     */
-    [[nodiscard]] auto getObsName() const -> const std::vector<std::string> & {
-        return obs_name_;
-    }
-    /**
-     * @brief Get the parameters for the observable operations.
-     *
-     * @return const std::vector<std::vector<T>>&
-     */
-    [[nodiscard]] auto getObsParams() const
-        -> const std::vector<param_var_t> & {
-        return obs_params_;
-    }
-    /**
-     * @brief Get the wires for each observable operation.
-     *
-     * @return const std::vector<std::vector<size_t>>&
-     */
-    [[nodiscard]] auto getObsWires() const
-        -> const std::vector<std::vector<size_t>> & {
-        return obs_wires_;
-    }
-
-  private:
-    const std::vector<std::string> obs_name_;
-    const std::vector<param_var_t> obs_params_;
-    const std::vector<std::vector<size_t>> obs_wires_;
-};
-
-/**
- * @brief Utility class for encapsulating operations used by AdjointJacobian
- * class.
- *
- */
-template <class T> class OpsData {
-  private:
-    size_t num_par_ops_;
-    size_t num_nonpar_ops_;
-    const std::vector<std::string> ops_name_;
-    const std::vector<std::vector<T>> ops_params_;
-    const std::vector<std::vector<size_t>> ops_wires_;
-    const std::vector<bool> ops_inverses_;
-    const std::vector<std::vector<std::complex<T>>> ops_matrices_;
-
-  public:
-    /**
-     * @brief Construct an OpsData object, representing the serialized
-     * operations to apply upon the `%StateVector`.
-     *
-     * @param ops_name Name of each operation to apply.
-     * @param ops_params Parameters for a given operation ({} if optional).
-     * @param ops_wires Wires upon which to apply operation
-     * @param ops_inverses Value to represent whether given operation is
-     * adjoint.
-     * @param ops_matrices Numerical representation of given matrix if not
-     * supported.
-     */
-    OpsData(std::vector<std::string> ops_name,
-            const std::vector<std::vector<T>> &ops_params,
-            std::vector<std::vector<size_t>> ops_wires,
-            std::vector<bool> ops_inverses,
-            std::vector<std::vector<std::complex<T>>> ops_matrices)
-        : ops_name_{std::move(ops_name)}, ops_params_{ops_params},
-          ops_wires_{std::move(ops_wires)},
-          ops_inverses_{std::move(ops_inverses)}, ops_matrices_{
-                                                      std::move(ops_matrices)} {
-        num_par_ops_ = 0;
-        for (const auto &p : ops_params) {
-            if (!p.empty()) {
-                num_par_ops_++;
-            }
-        }
-        num_nonpar_ops_ = ops_params.size() - num_par_ops_;
-    };
-
-    /**
-     * @brief Construct an OpsData object, representing the serialized
-     operations to apply upon the `%StateVector`.
-     *
-     * @see  OpsData(const std::vector<std::string> &ops_name,
-            const std::vector<std::vector<T>> &ops_params,
-            const std::vector<std::vector<size_t>> &ops_wires,
-            const std::vector<bool> &ops_inverses,
-            const std::vector<std::vector<std::complex<T>>> &ops_matrices)
-     */
-    OpsData(const std::vector<std::string> &ops_name,
-            const std::vector<std::vector<T>> &ops_params,
-            std::vector<std::vector<size_t>> ops_wires,
-            std::vector<bool> ops_inverses)
-        : ops_name_{ops_name}, ops_params_{ops_params},
-          ops_wires_{std::move(ops_wires)}, ops_inverses_{std::move(
-                                                ops_inverses)},
-          ops_matrices_(ops_name.size()) {
-        num_par_ops_ = 0;
-        for (const auto &p : ops_params) {
-            if (p.size() > 0) {
-                num_par_ops_++;
-            }
-        }
-        num_nonpar_ops_ = ops_params.size() - num_par_ops_;
-    };
-
-    /**
-     * @brief Get the number of operations to be applied.
-     *
-     * @return size_t Number of operations.
-     */
-    [[nodiscard]] auto getSize() const -> size_t { return ops_name_.size(); }
-
-    /**
-     * @brief Get the names of the operations to be applied.
-     *
-     * @return const std::vector<std::string>&
-     */
-    [[nodiscard]] auto getOpsName() const -> const std::vector<std::string> & {
-        return ops_name_;
-    }
-    /**
-     * @brief Get the (optional) parameters for each operation. Given entries
-     * are empty ({}) if not required.
-     *
-     * @return const std::vector<std::vector<T>>&
-     */
-    [[nodiscard]] auto getOpsParams() const
-        -> const std::vector<std::vector<T>> & {
-        return ops_params_;
-    }
-    /**
-     * @brief Get the wires for each operation.
-     *
-     * @return const std::vector<std::vector<size_t>>&
-     */
-    [[nodiscard]] auto getOpsWires() const
-        -> const std::vector<std::vector<size_t>> & {
-        return ops_wires_;
-    }
-    /**
-     * @brief Get the adjoint flag for each operation.
-     *
-     * @return const std::vector<bool>&
-     */
-    [[nodiscard]] auto getOpsInverses() const -> const std::vector<bool> & {
-        return ops_inverses_;
-    }
-    /**
-     * @brief Get the numerical matrix for a given unsupported operation. Given
-     * entries are empty ({}) if not required.
-     *
-     * @return const std::vector<std::vector<std::complex<T>>>&
-     */
-    [[nodiscard]] auto getOpsMatrices() const
-        -> const std::vector<std::vector<std::complex<T>>> & {
-        return ops_matrices_;
-    }
-
-    /**
-     * @brief Notify if the operation at a given index is parametric.
-     *
-     * @param index Operation index.
-     * @return true Gate is parametric (has parameters).
-     * @return false Gate in non-parametric.
-     */
-    [[nodiscard]] inline auto hasParams(size_t index) const -> bool {
-        return !ops_params_[index].empty();
-    }
-
-    /**
-     * @brief Get the number of parametric operations.
-     *
-     * @return size_t
-     */
-    [[nodiscard]] auto getNumParOps() const -> size_t { return num_par_ops_; }
-
-    /**
-     * @brief Get the number of non-parametric ops.
-     *
-     * @return size_t
-     */
-    [[nodiscard]] auto getNumNonParOps() const -> size_t {
-        return num_nonpar_ops_;
-    }
-};
 
 /**
  * @brief Represent the logic for the adjoint Jacobian method of
@@ -600,27 +377,6 @@ template <class T = double> class AdjointJacobian {
     AdjointJacobian() = default;
 
     /**
-     * @brief Utility to create a given operations object.
-     *
-     * @param ops_name Name of operations.
-     * @param ops_params Parameters for each operation in ops_name.
-     * @param ops_wires Wires for each operation in ops_name.
-     * @param ops_inverses Indicate whether to take adjoint of each operation in
-     * ops_name.
-     * @param ops_matrices Matrix definition of an operation if unsupported.
-     * @return const OpsData<T>
-     */
-    auto createOpsData(
-        const std::vector<std::string> &ops_name,
-        const std::vector<std::vector<T>> &ops_params,
-        const std::vector<std::vector<size_t>> &ops_wires,
-        const std::vector<bool> &ops_inverses,
-        const std::vector<std::vector<std::complex<T>>> &ops_matrices = {{}})
-        -> OpsData<T> {
-        return {ops_name, ops_params, ops_wires, ops_inverses, ops_matrices};
-    }
-
-    /**
      * @brief Calculates the Jacobian for the statevector for the selected set
      * of parametric gates.
      *
@@ -631,96 +387,102 @@ template <class T = double> class AdjointJacobian {
      * gradient calculations given in `trainableParams`, and the overall number
      * of parameters for the gradient calculation provided within `num_params`.
      * The resulting row-major ordered `jac` matrix representation will be of
-     * size `trainableParams.size() * observables.size()`. OpenMP is used to
-     * enable independent operations to be offloaded to threads.
+     * size `jd.getSizeStateVec() * jd.getObservables().size()`. OpenMP is used
+     * to enable independent operations to be offloaded to threads.
      *
-     * @param psi Pointer to the statevector data.
-     * @param num_elements Length of the statevector data.
      * @param jac Preallocated vector for Jacobian data results.
-     * @param observables Observables for which to calculate Jacobian.
-     * @param operations Operations used to create given state.
-     * @param trainableParams List of parameters participating in Jacobian
-     * calculation.
-     * @param apply_operations Indicate whether to apply operations to psi prior
-     * to calculation.
+     * @param jd JacobianData represents the QuantumTape to differentiate
+     * @param apply_operations Indicate whether to apply operations to tape.psi
+     * prior to calculation.
      */
-    void adjointJacobian(const std::complex<T> *psi, size_t num_elements,
-                         std::vector<std::vector<T>> &jac,
-                         const std::vector<ObsDatum<T>> &observables,
-                         const OpsData<T> &operations,
-                         const std::vector<size_t> &trainableParams,
+    void adjointJacobian(std::vector<T> &jac, const JacobianData<T> &jd,
                          bool apply_operations = false) {
-        PL_ABORT_IF(trainableParams.empty(),
+        PL_ABORT_IF(!jd.hasTrainableParams(),
                     "No trainable parameters provided.");
 
-        // Track positions within par and non-par operations
-        size_t num_observables = observables.size();
-        size_t trainableParamNumber = trainableParams.size() - 1;
-        size_t current_param_idx =
-            operations.getNumParOps() - 1; // total number of parametric ops
+        const OpsData<T> &ops = jd.getOperations();
+        const std::vector<std::string> &ops_name = ops.getOpsName();
 
-        auto tp_it = trainableParams.end();
+        const std::vector<ObsDatum<T>> &obs = jd.getObservables();
+        const size_t num_observables = obs.size();
+
+        const std::vector<size_t> &tp = jd.getTrainableParams();
+        const size_t tp_size = tp.size();
+        const size_t num_param_ops = ops.getNumParOps();
+
+        // Track positions within par and non-par operations
+        size_t trainableParamNumber = tp_size - 1;
+        size_t current_param_idx =
+            num_param_ops - 1; // total number of parametric ops
 
         // Create $U_{1:p}\vert \lambda \rangle$
-        StateVectorManaged<T> lambda(psi, num_elements);
+        StateVectorManaged<T> lambda(jd.getPtrStateVec(), jd.getSizeStateVec());
 
         // Apply given operations to statevector if requested
         if (apply_operations) {
-            applyOperations(lambda, operations);
+            applyOperations(lambda, ops);
         }
+
+        const auto tp_begin = tp.begin();
+        auto tp_it = tp.end();
 
         // Create observable-applied state-vectors
         std::vector<StateVectorManaged<T>> H_lambda(
             num_observables, StateVectorManaged<T>{lambda.getNumQubits()});
-        applyObservables(H_lambda, lambda, observables);
+        applyObservables(H_lambda, lambda, obs);
 
         StateVectorManaged<T> mu(lambda.getNumQubits());
 
-        for (int op_idx = static_cast<int>(operations.getOpsName().size() - 1);
-             op_idx >= 0; op_idx--) {
-            PL_ABORT_IF(operations.getOpsParams()[op_idx].size() > 1,
+        for (int op_idx = static_cast<int>(ops_name.size() - 1); op_idx >= 0;
+             op_idx--) {
+            PL_ABORT_IF(ops.getOpsParams()[op_idx].size() > 1,
                         "The operation is not supported using the adjoint "
                         "differentiation method");
-            if ((operations.getOpsName()[op_idx] != "QubitStateVector") &&
-                (operations.getOpsName()[op_idx] != "BasisState")) {
+            if ((ops_name[op_idx] != "QubitStateVector") &&
+                (ops_name[op_idx] != "BasisState")) {
                 mu.updateData(lambda.getDataVector());
-                applyOperationAdj(lambda, operations, op_idx);
+                applyOperationAdj(lambda, ops, op_idx);
 
-                if (operations.hasParams(op_idx)) {
-                    if (std::find(trainableParams.begin(), tp_it,
-                                  current_param_idx) != tp_it) {
+                if (ops.hasParams(op_idx)) {
+                    if ((current_param_idx == *(std::prev(tp_it))) ||
+                        std::find(tp_begin, tp_it, current_param_idx) !=
+                            tp_it) {
                         const T scalingFactor =
-                            applyGenerator(
-                                mu, operations.getOpsName()[op_idx],
-                                operations.getOpsWires()[op_idx],
-                                !operations.getOpsInverses()[op_idx]) *
-                            (2 * (operations.getOpsInverses()[op_idx] ? 0 : 1) -
-                             1);
+                            applyGenerator(mu, ops_name[op_idx],
+                                           ops.getOpsWires()[op_idx],
+                                           !ops.getOpsInverses()[op_idx]) *
+                            (ops.getOpsInverses()[op_idx] ? -1 : 1);
+
+                        const size_t mat_row_idx =
+                            trainableParamNumber * num_observables;
+
                         // clang-format off
 
                         #if defined(_OPENMP)
                             #pragma omp parallel for default(none)   \
                             shared(H_lambda, jac, mu, scalingFactor, \
-                                trainableParamNumber, tp_it,         \
+                                mat_row_idx,        \
                                 num_observables)
                         #endif
 
                         // clang-format on
                         for (size_t obs_idx = 0; obs_idx < num_observables;
                              obs_idx++) {
-                            updateJacobian(H_lambda[obs_idx], mu, jac,
-                                           scalingFactor, obs_idx,
-                                           trainableParamNumber);
+                            jac[mat_row_idx + obs_idx] =
+                                -2 * scalingFactor *
+                                std::imag(innerProdC(
+                                    H_lambda[obs_idx].getDataVector(),
+                                    mu.getDataVector()));
                         }
                         trainableParamNumber--;
                         std::advance(tp_it, -1);
                     }
                     current_param_idx--;
                 }
-                applyOperationsAdj(H_lambda, operations,
-                                   static_cast<size_t>(op_idx));
+                applyOperationsAdj(H_lambda, ops, static_cast<size_t>(op_idx));
             }
         }
+        jac = Transpose(jac, jd.getNumParams(), num_observables);
     }
 }; // class AdjointJacobian