+++ pennylane_lightning/src/bindings/Bindings.hpp
@@ -20,8 +20,11 @@
 #include "AdjointDiff.hpp"
 #include "IndicesUtil.hpp"
 #include "JacobianProd.hpp"
+#include "Measures.hpp"
 #include "StateVectorRaw.hpp"
+
 #include "pybind11/complex.h"
+#include "pybind11/functional.h"
 #include "pybind11/numpy.h"
 #include "pybind11/pybind11.h"
 #include "pybind11/stl.h"