+++ pennylane_lightning/src/simulator/Measures.hpp
@@ -25,12 +25,11 @@
 #include <cstdio>
 #include <vector>
 
+#include "LinearAlgebra.hpp"
 #include "StateVectorManaged.hpp"
 #include "StateVectorRaw.hpp"
-#include "Util.hpp"
 
 namespace Pennylane {
-using namespace Util;
 /**
  * @brief Observable's Measurement Class.
  *
@@ -70,15 +69,16 @@ class Measures {
     /**
      * @brief Probabilities for a subset of the full system.
      *
-     * @param wires Wires will restrict probabilities to a subset of the full
-     * system.
+     * @param wires Wires will restrict probabilities to a subset
+     * of the full system.
      * @return Floating point std::vector with probabilities.
      * The basis columns are rearranged according to wires.
      */
     std::vector<fp_t> probs(const std::vector<size_t> &wires) {
         // Determining index that would sort the vector.
         // This information is needed later.
-        const std::vector<size_t> sorted_ind_wires(sorting_indices(wires));
+        const std::vector<size_t> sorted_ind_wires(
+            Util::sorting_indices(wires));
         // Sorting wires.
         std::vector<size_t> sorted_wires(wires.size());
         for (size_t pos = 0; pos < wires.size(); pos++) {
@@ -109,14 +109,14 @@ class Measures {
         // the begining.
         if (wires != sorted_wires) {
             probabilities =
-                transpose_state_tensor(probabilities, sorted_ind_wires);
+                Util::transpose_state_tensor(probabilities, sorted_ind_wires);
         }
         return probabilities;
     }
     /**
      * @brief Expected value of an observable.
      *
-     * @param operation Square matrix in row-major order.
+     * @param matrix Square matrix in row-major order.
      * @param wires Wires where to apply the operator.
      * @return Floating point expected value of the observable.
      */
@@ -128,9 +128,9 @@ class Measures {
 
         operator_statevector.applyMatrix(matrix, wires);
 
-        CFP_t expected_value = innerProdC(original_statevector.getData(),
-                                          operator_statevector.getData(),
-                                          original_statevector.getLength());
+        CFP_t expected_value = Util::innerProdC(
+            original_statevector.getData(), operator_statevector.getData(),
+            original_statevector.getLength());
         return std::real(expected_value);
     };
     /**
@@ -148,9 +148,9 @@ class Measures {
 
         operator_statevector.applyOperation(operation, wires);
 
-        CFP_t expected_value = innerProdC(original_statevector.getData(),
-                                          operator_statevector.getData(),
-                                          original_statevector.getLength());
+        CFP_t expected_value = Util::innerProdC(
+            original_statevector.getData(), operator_statevector.getData(),
+            original_statevector.getLength());
         return std::real(expected_value);
     };
 
@@ -158,7 +158,6 @@ class Measures {
      * @brief Expected value for a list of observables.
      *
      * @param operations_list List of operations to measure.
-     * Square matrix in row-major order or string with the operator name.
      * @param wires_list List of wires where to apply the operators.
      * @return Floating point std::vector with expected values for the
      * observables.
@@ -185,8 +184,7 @@ class Measures {
     /**
      * @brief Variance of an observable.
      *
-     * @param operation Square matrix in row-major order or string with the
-     * operator name.
+     * @param operation String with the operator name.
      * @param wires Wires where to apply the operator.
      * @return Floating point with the variance of the observables.
      */
@@ -197,21 +195,21 @@ class Measures {
 
         operator_statevector.applyOperation(operation, wires);
 
-        fp_t mean_square = std::real(innerProdC(
-            operator_statevector.getData(), operator_statevector.getData(),
-            original_statevector.getLength()));
-        fp_t squared_mean = std::real(innerProdC(
-            original_statevector.getData(), operator_statevector.getData(),
-            original_statevector.getLength()));
-        squared_mean = std::pow(squared_mean, 2);
+        const std::complex<fp_t> *opsv_data = operator_statevector.getData();
+        size_t orgsv_len = original_statevector.getLength();
+
+        fp_t mean_square =
+            std::real(Util::innerProdC(opsv_data, opsv_data, orgsv_len));
+        fp_t squared_mean = std::real(Util::innerProdC(
+            original_statevector.getData(), opsv_data, orgsv_len));
+        squared_mean = static_cast<fp_t>(std::pow(squared_mean, 2));
         return (mean_square - squared_mean);
     };
 
     /**
      * @brief Variance of an observable.
      *
-     * @param operation Square matrix in row-major order or string with the
-     * operator name.
+     * @param operation Square matrix in row-major order.
      * @param wires Wires where to apply the operator.
      * @return Floating point with the variance of the observables.
      */
@@ -223,13 +221,14 @@ class Measures {
 
         operator_statevector.applyMatrix(matrix, wires);
 
-        fp_t mean_square = std::real(innerProdC(
-            operator_statevector.getData(), operator_statevector.getData(),
-            original_statevector.getLength()));
-        fp_t squared_mean = std::real(innerProdC(
-            original_statevector.getData(), operator_statevector.getData(),
-            original_statevector.getLength()));
-        squared_mean = std::pow(squared_mean, 2);
+        const std::complex<fp_t> *opsv_data = operator_statevector.getData();
+        size_t orgsv_len = original_statevector.getLength();
+
+        fp_t mean_square =
+            std::real(Util::innerProdC(opsv_data, opsv_data, orgsv_len));
+        fp_t squared_mean = std::real(Util::innerProdC(
+            original_statevector.getData(), opsv_data, orgsv_len));
+        squared_mean = static_cast<fp_t>(std::pow(squared_mean, 2));
         return (mean_square - squared_mean);
     };
 
@@ -259,5 +258,4 @@ class Measures {
         return expected_value_list;
     };
 }; // class Measures
-
 } // namespace Pennylane