+++ pennylane_lightning/src/simulator/SelectGateOps.hpp
@@ -17,6 +17,7 @@
  */
 #pragma once
 
+#include "ConstantUtil.hpp"
 #include "GateOperations.hpp"
 #include "GateOperationsLM.hpp"
 #include "GateOperationsPI.hpp"
@@ -336,7 +337,7 @@ std::vector<GateOperations> implementedGatesForKernelIter(KernelType kernel) {
 template <class fp_t>
 auto implementedGatesForKernel(KernelType kernel)
     -> std::vector<GateOperations> {
-    return Internal::implementedGatesForKernelIter<fp_t, 0>(kernel);
+    return implementedGatesForKernelIter<fp_t, 0>(kernel);
 }
 
 /********************************************************************