+++ .github/CHANGELOG.md
@@ -1,3 +1,27 @@
+# Release 0.22.0-dev
+
+### New features since last release
+
+### Breaking changes
+
+### Improvements
+
+* Set GitHub workflow to upload wheels to Test PyPI [(#220)](https://github.com/PennyLaneAI/pennylane-lightning/pull/220).
+
+### Documentation
+
+### Bug fixes
+
+* Refactor utility header and fix a bug in linear algebra function with CBLAS [(#228)](https://github.com/PennyLaneAI/pennylane-lightning/pull/228).
+
+### Contributors
+
+This release contains contributions from (in alphabetical order):
+
+Chae-Yeun Park
+
+---
+
 # Release 0.21.0
 
 ### New features since last release
@@ -10,6 +34,9 @@
 
 * C++ support for probability, expectation value and variance calculations.
 [(#185)](https://github.com/PennyLaneAI/pennylane-lightning/pull/185)
+
+* Add bindings to C++ expval, var, probs.
+[(#214)](https://github.com/PennyLaneAI/pennylane-lightning/pull/214)
 
 ### Improvements
 
@@ -48,6 +75,9 @@ The latest release shows improved performance on both single and multi-threaded 
 
 * Ensure debug info is built into dynamic libraries.
 [(#201)](https://github.com/PennyLaneAI/pennylane-lightning/pull/201)
+
+* Update adjointJacobian and VJP methods.
+[(#222)](https://github.com/PennyLaneAI/pennylane-lightning/pull/222)
 
 ### Documentation